extract JPEG boxes processing into its own file
