+++ libheif/box.cc
@@ -23,7 +23,7 @@
 #include "box.h"
 #include "security_limits.h"
 #include "nclx.h"
-#include "codecs/jpeg.h"
+#include "codecs/jpeg_boxes.h"
 #include "codecs/jpeg2000.h"
 #include "codecs/hevc_boxes.h"
 #include "codecs/mask_image.h"
