+++ libheif/codecs/jpeg.h
@@ -28,30 +28,6 @@
 #include <memory>
 
 
-class Box_jpgC : public Box
-{
-public:
-  Box_jpgC()
-  {
-    set_short_type(fourcc("jpgC"));
-  }
-
-  const std::vector<uint8_t>& get_data() const { return m_data; }
-
-  void set_data(const std::vector<uint8_t>& data) { m_data = data; }
-
-  std::string dump(Indent&) const override;
-
-  Error write(StreamWriter& writer) const override;
-
-protected:
-  Error parse(BitstreamRange& range) override;
-
-private:
-  std::vector<uint8_t> m_data;
-};
-
-
 class ImageItem_JPEG : public ImageItem
 {
 public:
